% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bind_rows2}
\alias{bind_rows2}
\title{A better (but slower) version of dplyr::bind_rows}
\usage{
bind_rows2(dfList, addMissing = TRUE, verbose = FALSE)
}
\arguments{
\item{dflist}{a list of data.frames}
}
\description{
A better (but slower) version of dplyr::bind_rows
}
