% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalSample.R
\name{condlSample.lm}
\alias{condlSample.lm}
\title{Generate samples from conditional distribution defined by a (generalized) linear model.}
\usage{
\method{condlSample}{lm}(object, newdata, quantile = "random", ...)
}
\arguments{
\item{object}{a lm object (including glm and gam). If a glm/gam, its family
must be gaussian, binomial, poisson, or gamma.}

\item{newdata}{As in predict.lm, predict.glm, etc. An optional data.frame to use for
generating conditional distribution parameters. If omitted, the fitted values are used.}

\item{quantile}{vector of quantiles to which the returned values will correspond.
If the default, 'random' is not used, this must be numeric on (0, 1).}

\item{...}{Passed to `predict()` method}
}
\value{
Numeric vector containing conditional random sample (if `quantile = 'random'`)
or conditional quantiles from condition distribution defined by `object` and `newdata`
}
\description{
Generate samples from conditional distribution defined by a (generalized) linear model.
}

