% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conditionalSample.R
\name{condlSample.rcgam}
\alias{condlSample.rcgam}
\title{Predict method for rcgam fits}
\usage{
\method{condlSample}{rcgam}(object, newdata, flowcol = "flow",
  flow.units = "CFS", quantile, ...)
}
\arguments{
\item{object}{An rcgam object to use for predicting}

\item{newdata}{a data.frame containing precictor variables to use for prediction}

\item{...}{Arguments passed to `predict.gam` function call}

\item{retransform}{Should the predictions be returned as concentrations? (defaults to TRUE)}

\item{smear}{Use Smearing estimator to correct transformation bias?}
}
\description{
Predict values using an rcgam model object
}

