% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalSample.R
\name{condlSample.rcmodel}
\alias{condlSample.rcmodel}
\title{Predict method for rclm/ rcgam fits}
\usage{
\method{condlSample}{rcmodel}(object, newdata, flowcol = "flow",
  flow.units = "CFS", quantile, ...)
}
\arguments{
\item{object}{An rclm or rcgam object to use for predicting}

\item{newdata}{a data.frame containing precictor variables to use for prediction}

\item{...}{Arguments passed to `predict.lm` or predict.gam function call}

\item{retransform}{Should the predictions be returned as concentrations? (defaults to TRUE)}

\item{smear}{Use Smearing estimator to correct transformation bias?}
}
\description{
Predict values using an rclm or rcgam model object
}

