% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{conflictsToCharacter}
\alias{conflictsToCharacter}
\title{Used in bind_rows2}
\usage{
conflictsToCharacter(df1, df2, addMissing = TRUE, verbose = FALSE)
}
\arguments{
\item{df1}{data.frame (or similar) to combine}

\item{df2}{data.frame (or similar) to combine}

\item{addMissing}{Add columns to result if they are missing in either df1 or df2?}

\item{verbose}{Include messages about conversions?}
}
\description{
Converts conflicted column types to character columns and rbinds the result. 
If one data.frame is missing columns, they are created and populated by NA's.
}
