% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidate.R
\name{crossvalidate.rcgam}
\alias{crossvalidate.rcgam}
\title{Crossvalidation for rcgam models}
\usage{
\method{crossvalidate}{rcgam}(object, kfolds = 0, statistic = c("R2", "mse",
  "mae", "rmse"), what = c("conc_mg.l", "load_kg.d"), retransform = TRUE,
  ...)
}
\arguments{
\item{what}{what quantity to predict/validate}

\item{retransform}{compare predictions to observations in original uints?
Must be TRUE if what == "load_kg.d"}

\item{...}{Passed to predict.rcgam}
}
\description{
Crossvalidation for rcgam models
}

