% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidate.R
\name{crossvalidate}
\alias{crossvalidate}
\title{Cross-validation of regression models}
\usage{
crossvalidate(object, ...)
}
\arguments{
\item{object}{a model object}

\item{kfolds}{Number of folds to use for crossvalidation. 0 (default) corresponds to leave-one-out crossvalidation.}

\item{statistic}{performance statistic to use.}
}
\description{
Generic function that computes a crossvalidation score for regression models.
}

