% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofStats.R
\name{delMSE}
\alias{delMSE}
\title{Deleted MSE
Useful for calculating deleted statistics like dffits, studentized deleted residuals}
\usage{
delMSE(model)
}
\arguments{
\item{model}{object, e.g. lm or gam}
}
\description{
Deleted MSE
Useful for calculating deleted statistics like dffits, studentized deleted residuals
}

