% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofStats.R
\name{delResid}
\alias{delResid}
\title{Deleted residuals, given a model object}
\usage{
delResid(object)
}
\arguments{
\item{object}{a model object that undertstands `residuals` and `hatvalues`}
}
\description{
Deleted residuals, given a model object
}

