% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfuns.R
\name{eigSample}
\alias{eigSample}
\title{Returns a data.frame with points sampled along first pincipal component of a covariance matrix
Only works for 2-D data}
\usage{
eigSample(sigma, n, quasi = c("none", "halton", "sobol"), ...)
}
\arguments{
\item{sigma}{a 2x2 covariance matrix}

\item{n}{number of points in sample}

\item{if}{not "none" (the default), uses a quasi-random sampling method}
}
\description{
Returns a data.frame with points sampled along first pincipal component of a covariance matrix
Only works for 2-D data
}

