% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crossvalidate.R
\name{gcv}
\alias{gcv}
\title{Generalized Crossvalidation score
Following Wood (2006), p. 178}
\usage{
gcv(object)
}
\arguments{
\item{object}{a model object with methods for `deviance`, `residuals` and `hatvalues`.}
}
\description{
Generalized Crossvalidation score
Following Wood (2006), p. 178
}

