% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Get data from lm objects}
\usage{
getData(object, ...)
}
\arguments{
\item{object}{an object inheriting from class `lm`}
}
\description{
simple extraction of data, returning useful errors if impossible. Useful as a 
method for the generic `getData` from `nlme` package
}
\seealso{
getData.rcgam
}
