% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfuns.R
\name{ggMvSample}
\alias{ggMvSample}
\title{Add points corresponding to a multivariate normal sample}
\usage{
ggMvSample(mvsample, ...)
}
\arguments{
\item{mvsample}{a data.frame returned by mvSample()}

\item{...}{Arguments passed to geom_point}
}
\description{
Add points corresponding to a multivariate normal sample
}
