% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfuns.R
\name{ggPcLines}
\alias{ggPcLines}
\title{Returns a geom_segment object with principal components shown}
\usage{
ggPcLines(pclines, ...)
}
\arguments{
\item{pclines}{a data.frame returned by pcLines function}

\item{...}{Arguments passed to geom_line}
}
\description{
Returns a geom_segment object with principal components shown
}

