% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggfuns.R
\name{ggTermPlot}
\alias{ggTermPlot}
\title{Termplot function for regression models
Currently only supports gam objects.}
\usage{
ggTermPlot(object, ...)
}
\arguments{
\item{...}{Passed to facet_wrap}
}
\description{
Termplot function for regression models
Currently only supports gam objects.
}

