% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfuns.R
\name{mvSample}
\alias{mvSample}
\title{Returns a data.frame with points sampled according to a covariance matrix
Only works for 2-D data}
\usage{
mvSample(sigma, n, quasi = c("none", "halton", "sobol"), ...)
}
\arguments{
\item{sigma}{a 2x2 covariance matrix}

\item{...}{other arguments passed to halton() or sobol() (if quasi != "none")}

\item{if}{not "none" (the default), uses a quasi-random sampling method}
}
\description{
Returns a data.frame with points sampled according to a covariance matrix
Only works for 2-D data
}

