% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidate.R
\name{ocv}
\alias{ocv}
\title{Ordinary Crossvalidation score
Following Wood (2006), p. 174}
\usage{
ocv(object)
}
\arguments{
\item{object}{a model object with methods for `residuals` and `hatvalues`.}
}
\description{
Ordinary Crossvalidation score
Following Wood (2006), p. 174
}
