% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfuns.R
\name{pcLines}
\alias{pcLines}
\title{Returns a data.frame with endpoints of principal components.
Only works for 2-D data}
\usage{
pcLines(sigma)
}
\arguments{
\item{pclines}{a data.frame returned by pcLines function}
}
\description{
Returns a data.frame with endpoints of principal components.
Only works for 2-D data
}
