% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofStats.R
\name{Q2}
\alias{Q2}
\title{Leave-one-out coefficient of determination
Similar to R2, but uses leave-one-out estimates for response and mean. 
Uses GCV and OCV to estimate MSE of prediction.}
\usage{
Q2(object, method = c("gcv", "ocv"))
}
\arguments{
\item{object}{an object with a class for `deviance`, `residuals`, and `hatvalues`.}
}
\description{
Leave-one-out coefficient of determination
Similar to R2, but uses leave-one-out estimates for response and mean. 
Uses GCV and OCV to estimate MSE of prediction.
}
