% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gofStats.R
\name{R2.default}
\alias{R2.default}
\title{Default S3 method for coefficient of determination}
\usage{
\method{R2}{default}(x, xpred)
}
\arguments{
\item{x}{observed data}

\item{xpred}{predictions}
}
\description{
Default S3 method for coefficient of determination
}

