% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofStats.R
\name{R2.gam}
\alias{R2.gam}
\title{Coefficient of dermination for gam objects
Optionally gives adjusted R2.}
\usage{
\method{R2}{gam}(x, adjust = FALSE)
}
\arguments{
\item{x}{an object of class "gam"}
}
\description{
Coefficient of dermination for gam objects
Optionally gives adjusted R2.
}
