% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofStats.R
\name{R2.lm}
\alias{R2.lm}
\title{Coefficient of dermination for lm objects
Optionally gives adjusted R2.}
\usage{
\method{R2}{lm}(x, adjust = FALSE)
}
\arguments{
\item{x}{an object of class "lm"}
}
\description{
Coefficient of dermination for lm objects
Optionally gives adjusted R2.
}

