% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartImpute.R
\name{randlr}
\alias{randlr}
\title{Impute BDL data using various distributions}
\usage{
randlr(lrdf, distr = c("unif", "norm", "lnorm"))
}
\arguments{
\item{lrdf}{data.frame with columns "left" and "right"}
}
\value{
vector with length nrow(lrdf) that contains original and imputed values
THIS VERSION RETURNS NA when distribution cannot be fit. An earlier version used uniform when this happened.
}
\description{
Function to impute BDL data, using either uniform or truncated lognormal distributions
The latter comes from the fitdistrplus::fitdistcens function
}

