% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfuns.R
\name{relDif}
\alias{relDif}
\title{Creates 2-d filled contour plot for relative differences between two 
data frames, each with 2 columns, "Date" and a numeric value.
Series 1 is reference, to which series2 is compared.
Value plotted is percent difference}
\usage{
relDif(series1, series2, log = T, breaks = NULL, cols = NULL)
}
\description{
Creates 2-d filled contour plot for relative differences between two 
data frames, each with 2 columns, "Date" and a numeric value.
Series 1 is reference, to which series2 is compared.
Value plotted is percent difference
}

