% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofStats.R
\name{rstudent.gam}
\alias{rstudent.gam}
\title{Studentized deleted residuals, given a model object}
\usage{
\method{rstudent}{gam}(object, ...)
}
\arguments{
\item{object}{a model object that undertstands `residuals` and `hatvalues`}
}
\description{
Studentized deleted residuals, given a model object
}

