% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfuns.R
\name{sampleToPCs}
\alias{sampleToPCs}
\title{Project observations from a sample to principal components, given a covariance matrix}
\usage{
sampleToPCs(x, sigma)
}
\arguments{
\item{x}{A data.frame, typically a (quasi-) random sample from MVN(0, I)}
}
\description{
Project observations from a sample to principal components, given a covariance matrix
}

