% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splitPiece}
\alias{splitPiece}
\title{Get all of the pieces for a particular stringsplit.}
\usage{
splitPiece(strvec, split, piece, ...)
}
\arguments{
\item{strvec}{Vector of strings, each to be split.}

\item{split}{The string to split on. See `help("strsplit")`}

\item{piece}{The integer location of the string split piece desired}

\item{...}{Other arguments passed to `strsplit()`}
}
\description{
Get all of the pieces for a particular stringsplit.
}
