% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggfuns.R
\name{subset.ggplot}
\alias{subset.ggplot}
\title{Subsets a ggplot object by applying subset.data.frame code to gg$data}
\usage{
\method{subset}{ggplot}(gg, subset, select, drop = FALSE)
}
\description{
Subsets a ggplot object by applying subset.data.frame code to gg$data
}

