% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{titleCase}
\alias{titleCase}
\title{Capitalize first letter of each word in text
Copied without shame from http://stackoverflow.com/a/15776815}
\usage{
titleCase(text)
}
\description{
Capitalize first letter of each word in text
Copied without shame from http://stackoverflow.com/a/15776815
}

