% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalSample.R
\name{condlSample}
\alias{condlSample}
\alias{condlSample.rcgam}
\alias{condlSample.rclm}
\title{Generate samples from conditional distribution.}
\usage{
condlSample(object, ...)

\method{condlSample}{rcgam}(object, newdata, flowcol = "flow",
  flow.units = "CFS", quantile, ...)

\method{condlSample}{rclm}(object, newdata, flowcol = "flow",
  flow.units = "CFS", quantile, ...)
}
\arguments{
\item{object}{a model object with a `predict()` method}

\item{...}{arguments to be passed to individual methods. These should always include :
`newdata` - to be passed to `predict()` function
`quantile` - either 'random' for stochastic sampling from conditional distribution or 
a numeric value on (0, 1) specifying the quantile to return}
}
\value{
Numeric vector containing conditional random sample (if `quantile = 'random'`)
or conditional quantiles from condition distribution defined by `object` and `newdata`
seealso \code{\link{condlSample.lm}}
}
\description{
Generate samples from conditional distribution.
}
\section{Methods (by class)}{
\itemize{
\item \code{rcgam}: Method for rcgam objects

\item \code{rclm}: Method for rclm objects
}}

