% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conditionalSample.R
\name{condlSample}
\alias{condlSample}
\title{Generate samples from conditional distribution.}
\usage{
condlSample(object, ...)
}
\arguments{
\item{object}{a model object with a `predict()` method}

\item{...}{arguments to be passed to individual methods. These should always include :
`newdata` - to be passed to `predict()` function
`quantile` - either 'random' for stochastic sampling from conditional distribution or
a numeric value on (0, 1) specifying the quantile to return}
}
\value{
Numeric vector containing conditional random sample (if `quantile = 'random'`)
or conditional quantiles from condition distribution defined by `object` and `newdata`
seealso \code{\link{condlSample.lm}}
}
\description{
Generate samples from conditional distribution.
}

