% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projTemplate.R
\name{fromcache}
\alias{fromcache}
\title{Fetch a single object from cache}
\usage{
fromcache(..., list = character(), envir = parent.frame())
}
\arguments{
\item{...}{the objects to be retrieved, as names (unquoted) or character strings (quoted).}

\item{list}{a character vector naming objects to be retrieved.}

\item{envir}{the environment into which to retrieve the object.}
}
\value{
None, used for side-effects only
}
\description{
Does the opposite of ProjectTemplate::cache -- given an object name (as a
character string), retrieves that object from the cache directory.
}
\seealso{
ProjectTemplate::cache
}
