% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projTemplate.R, R/utils.R
\name{fromcache}
\alias{fromcache}
\title{Fetch a single object from cache}
\usage{
fromcache(variable, envir = parent.frame())

fromcache(variable, envir = parent.frame())
}
\arguments{
\item{variable}{character, the name of the variable to be retrieved.}

\item{value}{None, used for side-effects only}

\item{variable}{character, the name of the variable to be retrieved.}

\item{value}{None, used for side-effects only}
}
\description{
Does the opposite of ProjectTemplate::cache -- given an object name (as a
character string), retrieves that object from the cache directory.

Does the opposite of ProjectTemplate::cache -- given an object name (as a
character string), retrieves that object from the cache directory.
}
\seealso{
ProjectTemplate::cache

ProjectTemplate::cache
}

