% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfuns.R
\name{ggTermPlot}
\alias{ggTermPlot}
\title{ggplot2 implementation of termplot with partial residuals, using visreg::visreg()}
\usage{
ggTermPlot(object, ...)
}
\arguments{
\item{object}{A model object that visreg() recognizes}

\item{xvar}{Which term to inspect. Currently only works for one at a time.}

\item{data}{If object$call$data isn't in the model environment, it can be supplied using this}
}
\description{
ggplot2 implementation of termplot with partial residuals, using visreg::visreg()
}
